+++ packages/config/src/old/loaders/cli.ts
@@ -1,6 +1,6 @@
-import type { ConfigLoader } from 'builder/base';
-import type { ConfigKeys } from 'loaders/base';
-import { LoaderInputError } from 'utils/errors';
+import type { ConfigLoader } from 'old/builder/base';
+import type { ConfigKeys } from 'old/loaders/base';
+import { LoaderInputError } from 'old/utils/errors';
 
 export interface CLILoader {
   prefix: string;
