+++ packages/config/src/old/utils/errors.ts
