+++ packages/config/src/old/utils/freeze.ts
