+++ packages/config/src/old/utils/translators/conventions.ts
